mv stop reason norm to ext utils
