+++ python/packages/autogen-core/src/autogen_core/models/__init__.py
@@ -3,7 +3,6 @@ from ._model_client import (
     ModelCapabilities,  # type: ignore
     ModelFamily,
     ModelInfo,
-    normalize_stop_reason,
     validate_model_info,
 )
 from ._types import (
@@ -36,6 +35,5 @@ __all__ = [
     "ChatCompletionTokenLogprob",
     "ModelFamily",
     "ModelInfo",
-    "normalize_stop_reason",
     "validate_model_info",
 ]
